rename slim module to `linear`, move PopularityRecommender
